career-guidance-system/backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
openai==1.75.0
sentence-transformers==3.4.1
transformers==4.49.0
torch>=2.2.0
faiss-cpu==1.10.0
numpy==2.2.3
spacy==3.8.4
pypdf==5.4.0
